requirements.txt
streamlit
streamlit_chat
langchain
openai
wikipedia
chromadb
tiktoken
pypdf
unstructured[local-inference]
python-dotenv
huggingface_hub
InstructorEmbedding
sentence-transformers